Added Total Ram

+++ src/code5/stdlib.c
@@ -1,6 +1,6 @@
-void itoa(uint32_t n, uint32_t* buffer, uint32_t basenumber)
+void itoa(uint64_t n, uint64_t* buffer, uint32_t basenumber)
 {
-	uint32_t hold;
+	uint64_t hold;
 	int16_t i, j;
 	hold = n;
 	i = 0;
