+++ src/code5/stdlib.h
@@ -4,6 +4,6 @@
 #ifndef STDLIB_H
 #define STDLIB_H
 
-void itoa(uint32_t n, uint32_t* buffer, uint32_t basenumber);
+void itoa(uint64_t n, uint64_t* buffer, uint32_t basenumber);
 
 #endif // STDLIB_H
